[FIX] sometimes description got stuck
git-svn-id: https://scm.arca.ime.usp.br/morcego/trunk@84 b160122e-fef4-0310-8144-9a7482f4e04c

+++ src/main/java/br/arca/morcego/Config.java
@@ -154,7 +154,7 @@ public class Config {
 
 		setValue(minNodeSize, new Integer(0));
 		setValue(navigationDepth, new Integer(3));
-		setValue(feedAnimationInterval, new Integer(200));
+		setValue(feedAnimationInterval, new Integer(100));
 		setValue(balancingStepInterval, new Integer(50));
 		setValue(fontSizeInterval, new Integer(5));
